step_136-Done

+++ student_revinskaya_ekaterina_version_2/src/main/java/lv/javaguru/travel/insurance/core/repositories/TravelCostCoefficientRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public interface TravelCostCoefficientRepository extends JpaRepository<TravelCostCoefficient, Long> {
@@ -13,6 +14,6 @@ public interface TravelCostCoefficientRepository extends JpaRepository<TravelCos
             "where tc.costFrom <= :cost " +
             "and tc.costTo > :cost")
     Optional<TravelCostCoefficient> findByCost(
-            @Param("cost") int cost
+            @Param("cost") BigDecimal cost
     );
 }
